Fix issues with eslint-bulk prune.
