+++ common/changes/@rushstack/eslint-patch/fix-bulk-suppression-prune_2024-03-27-19-05.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@rushstack/eslint-patch",
+      "comment": "Fix an issue where `eslint-bulk prune` does not work if there are no files to lint in the project root.",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@rushstack/eslint-patch"
+}
